% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cs_offset.R
\name{add_cs_offset}
\alias{add_cs_offset}
\title{Chemical Shift Referencing Function}
\usage{
add_cs_offset(cs_input, cs_offset)
}
\arguments{
\item{cs_input}{chemical shift dataframe. Should contain field: resname, resid, nucleus, expCS.}

\item{cs_offset}{chemical shift offset dataframe. Should contain field: resname, resid, nucleus, offset.}
}
\description{
This function allows you to add an offset to chemical shift based on nucleus type and resname (i.e., residue name)
}
\examples{
add_cs_offset(cs, offsets)
}

