% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cs_weights.R
\name{add_cs_weights}
\alias{add_cs_weights}
\title{Differential Weighting Function}
\usage{
add_cs_weights(cs_input, weights, atomBasedWeights = FALSE)
}
\arguments{
\item{cs_input}{chemical shift dataframe. Should contain field: resname, resid, nucleus, expCS.}

\item{weights}{chemical shift weights dataframe. Should contain field: nucleus and weight.}

\item{atomBasedWeights}{should weights by added based on nucleus type and resname. If TRUE, weights dataframe should also contain a resname field. Defaults to FALSE.}
}
\description{
This function allows you to add weights chemical shift data
}
\examples{
add_cs_weights(cs, weights, atomBasedWeights=FALSE)
add_cs_weights(cs, weights, atomBasedWeights=TRUE)
}
