% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_2D_peaks.R
\name{compass_score}
\alias{compass_score}
\title{COMPASS Scoring Function}
\usage{
compass_score(Q, P)
}
\arguments{
\item{Q}{reference spectrum (experimental)}

\item{P}{comparison spectrum (simulated)}
}
\description{
This function allows you to compare two multi-dimensional spectra
See: 10.1016/j.str.2015.07.019: Experimental Protein Structure Verification by Scoring with a Single, Unassigned NMR Spectrum
}
\examples{
compass_score(cs)
}
