% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{correlation_kendall}
\alias{correlation_kendall}
\title{Kendall Correlation Scoring Function}
\usage{
correlation_kendall(x)
}
\arguments{
\item{x}{input dataframe. Should contain field: predCS and expCS.}
}
\description{
This function computes the 1 - tau
}
\examples{
score_kendall(x)
}

