% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{correlation_pearson}
\alias{correlation_pearson}
\title{Pearson Correlation Scoring Function}
\usage{
correlation_pearson(x)
}
\arguments{
\item{x}{input dataframe. Should contain field: predCS and expCS.}
}
\description{
This function computes the 1 - R
}
\examples{
score_pearson(x)
}
