% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_1D_peaks.R
\name{create_1D_specturm}
\alias{create_1D_specturm}
\title{1D Spectrum Generation Function}
\usage{
create_1D_specturm(cs, bw = 0.01, proton = TRUE)
}
\arguments{
\item{cs}{input chemical shift dataframe. Should contain field: model, resid, nucleus, weight, and predCS}

\item{bw}{bandwidth for density calculation}

\item{proton}{generate spectrum for proton nuclei; if not, will generate spectrum for carbon.}
}
\description{
This function generate a 1D NMR spectrum for chemical shift
}
\examples{
create_1D_specturm(cs, proton = FALSE)
}
