% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cs_hmqc_peaks.R
\name{create_peaks}
\alias{create_peaks}
\title{HMQC Peaks Generation Function}
\usage{
create_peaks(cs, protons = c("H1'", "H2'", "H3'", "H4'", "H5'", "H5''", "H2",
  "H5", "H6", "H8"), carbons = c("C1'", "C2'", "C3'", "C4'", "C5'", "C5'",
  "C2", "C5", "C6", "C8"), grouping = c("model", "resid"))
}
\arguments{
\item{cs}{input chemical shift dataframe. Should contain field: model, resid, nucleus, weight, and predCS}

\item{protons}{vector of proton nuclei.}

\item{carbons}{vector of carbon nuclei.}

\item{grouping}{variables used to group data}
}
\description{
This function allows you to convert chemical shift list to chemical shift peak table
}
\examples{
create_peaks(cs)
}
