% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_2D_peaks.R
\name{create_residue_tocsy_peaks}
\alias{create_residue_tocsy_peaks}
\title{TOCSY Peaks Generation Helper Function}
\usage{
create_residue_tocsy_peaks(cs)
}
\arguments{
\item{cs}{input chemical shift dataframe. Should contain field: model, resid, nucleus, weight, and predCS}
}
\description{
This function allows you to convert chemical shift list to chemical shift peak table
}
\examples{
create_residue_tocsy_peaks(cs)
}
