% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_referencing_error.R
\name{detect_referencing_error}
\alias{detect_referencing_error}
\title{Automated Referencing Error Detection Using Bayesian Regression}
\usage{
detect_referencing_error(observed_chemical_shifts, computed_chemical_shifts,
  residue_and_nucleus_based = FALSE, verbose = FALSE)
}
\arguments{
\item{observed_chemical_shifts}{observed chemical shift dataframe. Should contain field: resname, resid, nucleus, expCS.}

\item{computed_chemical_shifts}{observed chemical shift dataframe. Should contain field: resname, resid, nucleus, predCS}

\item{residue_and_nucleus_based}{if TRUE, both residue type and nucleus are considered when detecting errors. Default considers only nucleus type}

\item{verbose}{if TRUE, print progress log from MCMCpack. Default is FALSE}
}
\description{
This function allows user to detect potential referencing errors by comparing observed and computed chemical shifts
}
\examples{
detect_referencing_error(observed_chemical_shifts, computed_chemical_shifts, residue_and_nucleus_based=TRUE)
detect_referencing_error(observed_chemical_shifts, computed_chemical_shifts, residue_and_nucleus_based=FALSE)
}

