% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_bmrb.R
\name{download_and_parse_bmrb}
\alias{download_and_parse_bmrb}
\title{Download and Parse Function}
\usage{
download_and_parse_bmrb(ID, verbose = FALSE)
}
\arguments{
\item{ID}{BMRB database ID}

\item{verbose}{logical indicating print info to screen}
}
\description{
download the BMRB file specified by ID and generate a chemical shift data frame
}
