% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_factor.R
\name{enrichment_factor}
\alias{enrichment_factor}
\title{Enrichment Factor Function}
\usage{
enrichment_factor(X, fraction = 0.05)
}
\arguments{
\item{X}{vector of 0 (inactives) and 1 (actives) that was sorted based on some scores (e.g., agreement between measured and predicted shifts)}
}
\description{
This function computes the enrichment factor
}
\examples{
enrichment_factor(sample(c(rep(0,100),rep(1,10))))
}
