% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{geometric.mean}
\alias{geometric.mean}
\title{Geometric Mean Function}
\usage{
geometric.mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{input dataframe.}
}
\description{
The geometric mean is the nth root of n products or e to the mean log of x. (Copied from the "psych".)
}
\examples{
geometric.mean(x)
}
