% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_bmrb.R
\name{get_bmrb_list_info}
\alias{get_bmrb_list_info}
\title{Count Datasets Function}
\usage{
get_bmrb_list_info(y)
}
\arguments{
\item{y}{dataframe returned by download_and_parse_bmrb}
}
\description{
counting chemcial shift lists
}

