% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_bmrb.R
\name{get_bmrb_summary}
\alias{get_bmrb_summary}
\title{Summary Information Function}
\usage{
get_bmrb_summary(ID, y)
}
\arguments{
\item{ID}{BMRB database ID}

\item{y}{dataframe returned by download_and_parse_bmrb}
}
\description{
prints of general info about the BMRB entry (e.g., temperature and pH)
}

