% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cs_subset.R
\name{get_cs_subset}
\alias{get_cs_subset}
\title{A Chemical Shift Selection Function}
\usage{
get_cs_subset(cs_input, nuc = "all")
}
\arguments{
\item{cs_input}{chemical shift dataframe}

\item{nuc}{name of nucleus type to select}
}
\description{
This function allows you to select a subset of the chemical shift data based on nucleus type
}
\examples{
get_cs_subset(cs, "H1'")  
get_cs_subset(cs, "all")  
get_cs_subset(cs, "baseCarbon")  
}
