% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{get_ensemble_average}
\alias{get_ensemble_average}
\title{Ensemble Averaging Function}
\usage{
get_ensemble_average(w, X)
}
\arguments{
\item{w}{weights vector}

\item{X}{data matrix to be averaged}
}
\description{
This function compute ensemble-averaged of X using w
}
\examples{
get_ensemble_average(w, X)  
}
\author{
Aaron T. Frank
}
