% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{kendall_score}
\alias{kendall_score}
\title{Kendall Correlation Scoring Function}
\usage{
kendall_score(x)
}
\arguments{
\item{x}{input dataframe. Should contain field: predCS and expCS.}
}
\description{
This function computes the 1 - tau
}
\examples{
kendall_score(x)
}

