% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cs_data.R
\name{load_cs_data}
\alias{load_cs_data}
\title{Chemical Shift Data Loading Function}
\usage{
load_cs_data(csfile, remove_assigned = FALSE, accuracyFile = NULL,
  cs_offset = NULL, residues = NULL, atomBasedWeights = FALSE,
  names = c("processor", "model", "resid", "resname", "nucleus", "predCS",
  "expCS", "randCS", "id"))
}
\arguments{
\item{csfile}{input file name}

\item{remove_assigned}{remove data with zero values? Defaults to FALSE}

\item{accuracyFile}{expected accuracy file. Defaults to NULL}

\item{cs_offset}{file containing offsets. Defaults to NULL}

\item{residues}{vector of residues to retain data for. Defaults to NULL}

\item{atomBasedWeights}{vector of specifying the residues for to data will be retained. Defaults to NULL}

\item{names}{vector of column names. Defaults to c("processor","model","resid","resname","nucleus","predCS","expCS","randCS","id")}
}
\description{
This function allows you to load chemical shift data from file
}
\examples{
load_cs_data("data/predicted_shifts.txt")
}
