% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cs_data.R
\name{load_cs_matrix}
\alias{load_cs_matrix}
\title{Chemical Shift Matrix Loading Function}
\usage{
load_cs_matrix(csfile, skip = 1, cs_offset = NULL, residues = NULL,
  names = c("resid", "resname", "C1'", "C2'", "C3'", "C4'", "C5'", "C2", "C5",
  "C6", "C8", "H1'", "H2'", "H3'", "H4'", "H2", "H5", "H5'", "H5''", "H6", "H8",
  "H1", "H3", "N1", "N3"))
}
\arguments{
\item{csfile}{input file name}

\item{cs_offset}{file containing offsets. Defaults to NULL}

\item{residues}{vector of residues to retain data for. Defaults to NULL}

\item{names}{vector of column names. Defaults to c("processor","model","resid","resname","nucleus","predCS","expCS","randCS","id")}

\item{accuracyFile}{expected accuracy file. Defaults to NULL}

\item{atomBasedWeights}{vector of specifying the residues for to data will be retained. Defaults to NULL}
}
\description{
This function allows you to load chemical shift data from file in the form of a matrix
}
\examples{
load_cs_matrix("data/predicted_shifts.txt")
}
