% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_library.R
\name{load_library}
\alias{load_library}
\title{Load Package Function}
\usage{
load_library(package)
}
\arguments{
\item{package}{name of package to load.}
}
\description{
This function loads a package if previous installed or install the package and then loads it.
}

