% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_errors.R
\name{model_erorrs}
\alias{model_erorrs}
\title{Model Errors Function}
\usage{
model_erorrs(cs_input, errorType)
}
\arguments{
\item{cs_input}{chemical shift dataframe. Should contain field: resname, resid, nucleus, expCS, weight.}

\item{n_total}{total number of samples. Defaults to 1000}

\item{cycles}{number of random ranking to generate. Default to 10000.}
}
\description{
This function allows you to computes the error for each indiviudal model (or conformer)
}
\examples{
random_nslr(10, n_total = 100, cycles = 1000)
}
