% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{pearson_score}
\alias{pearson_score}
\title{Pearson Correlation Scoring Function}
\usage{
pearson_score(x)
}
\arguments{
\item{x}{input dataframe. Should contain field: predCS and expCS.}
}
\description{
This function computes the 1 - R
}
\examples{
pearson_score(x)
}

