% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_nslr.R
\name{random_nslr}
\alias{random_nslr}
\title{Randomized Sum of Logarithmic Ranks Function}
\usage{
random_nslr(n_actives, n_total = 1000, cycles = 10000)
}
\arguments{
\item{n_actives}{number of positive samples.}

\item{n_total}{total number of samples. Defaults to 1000}

\item{cycles}{number of random ranking to generate. Default to 10000.}
}
\description{
This function allows you to compute a randomize normalized sum of logarithmic ranks
}
\examples{
random_nslr(10, n_total = 100, cycles = 1000)
}

