% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_bmrb.R
\name{referencing_errors_detector}
\alias{referencing_errors_detector}
\title{Referencing Error Detection Function}
\usage{
referencing_errors_detector(input, first = NULL, last = NULL)
}
\arguments{
\item{input}{chemical shift data frame (expects columns to correspond to "resname","resid","nucleus","expCS","error" or "resname","resid","nucleus","expCS"}

\item{first:}{integer specifying residue number of 5' terminal residue in chemical shift list file}

\item{last:}{integer specifying residue number of 3' terminal residue in chemical shift list file}
}
\description{
Identifies whether 13C chemical shift data is systematically shifted. Use the chemical shifts of the a terminal G:C base pair and compare to refernce values
}
