% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_factor.R
\name{relative_enrichment_factor}
\alias{relative_enrichment_factor}
\title{Relative Enrichment Factor Function}
\usage{
relative_enrichment_factor(X, fraction = 0.1)
}
\arguments{
\item{X}{vector of 0 (inactives) and 1 (actives) that was sorted based on some scores (e.g., agreement between measured and predicted shifts)}
}
\description{
This function computes the relative enrichment factor
}
\examples{
relative_enrichment_factor(sample(c(rep(0,100),rep(1,10))))
}
