% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{run_ga_selection}
\alias{run_ga_selection}
\title{GA Optimization Function}
\usage{
run_ga_selection(target, ensemble, cycles = 100, population_size = 100,
  seed = 12345, binary = TRUE, weights = 1, mask = NULL,
  monitor = FALSE, alpha = 10)
}
\arguments{
\item{target}{Actual data (vector)}

\item{ensemble}{Matrix of predicted data. Number of columns correspond to the number of ensemble members and rows the number sample points in the target data.}

\item{cycles}{GA optimization cycles (integer)}

\item{population_size}{GA population size (integer)}

\item{seed}{Seed for random number generator}

\item{binary}{Run binary selection mode?}

\item{weights}{Weights used to individual data points}

\item{monitor}{Print progress?}

\item{alpha}{Coefficient to L1 regularization term}
}
\description{
This function runs GA optimizations
}
\examples{
fitness(p)  
}
\author{
Aaron T. Frank
}
