% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{score_flat_chi2}
\alias{score_flat_chi2}
\title{Weighted Flat Chi^2 Error Function}
\usage{
score_flat_chi2(x, chi2_c = 1)
}
\arguments{
\item{x}{input dataframe. Should contain field: weight, predCS and expCS.}

\item{chi2_c}{double. The scaling factor for the error in the exponent.}
}
\description{
This function computes  the (or reduced) flat Chi^2 error (return 1 - flat_chi2).
}
\examples{
score_flat_chi2(x)
}

