% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{score_geo_mae}
\alias{score_geo_mae}
\title{Weighted Mean Absolute Error Function}
\usage{
score_geo_mae(x)
}
\arguments{
\item{x}{input dataframe. Should contain field: weight, predCS and expCS.}
}
\description{
This function computes the weighted (or reduced) mean-absolute-error (MAE)
}
\examples{
score_geo_mae(x)
}

