% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{score_huber}
\alias{score_huber}
\title{Pseudo-Huber Error Function}
\usage{
score_huber(x)
}
\arguments{
\item{x}{input dataframe. Should contain field: w, predCS and expCS.}
}
\description{
This function a weighted (or reduced) root-mean-squared-error (RMSE)
}
\examples{
score_rmse(x)
}

