% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{score_rmse}
\alias{score_rmse}
\title{Weighted Root Mean Squared Error Function}
\usage{
score_rmse(x)
}
\arguments{
\item{x}{input dataframe. Should contain field: weight, predCS and expCS.}
}
\description{
This function computes the weighted (or reduced) root-mean-squared-error (RMSE)
}
\examples{
score_rmse(x)
}
