% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{score_spearman}
\alias{score_spearman}
\title{Spearman Correlation Scoring Function}
\usage{
score_spearman(x)
}
\arguments{
\item{x}{input dataframe. Should contain field: predCS and expCS.}
}
\description{
This function computes the 1 - rho
}
\examples{
score_spearman(x)
}
