% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{spearman_score}
\alias{spearman_score}
\title{Spearman Correlation Scoring Function}
\usage{
spearman_score(x)
}
\arguments{
\item{x}{input dataframe. Should contain field: predCS and expCS.}
}
\description{
This function computes the 1 - rho
}
\examples{
spearman_score(x)
}

