% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{update_package}
\alias{update_package}
\title{Packahe Documentation Update Function}
\usage{
update_package(package_path = "/Users/atfrank/GitSoftware/nmR",
  package_name = "nmR")
}
\arguments{
\item{package_path}{path (string) to package}

\item{package_name}{name (string) of package}
}
\description{
This function updates the package specified by package_path and package_name
}
\examples{
update_package(package_path='/Users/atfrank/GitSoftware/nmR', package_name='nmR')
}

