% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_bmrb.R
\name{write_out_shifts}
\alias{write_out_shifts}
\title{Chemical Shift Output Function}
\usage{
write_out_shifts(y, rnaOnly = TRUE, listID = NULL, offset = offset,
  rnaID = "1234", outpath = "", residues = NULL, nuclei = NULL,
  str = FALSE, verbose = TRUE)
}
\arguments{
\item{y}{dataframe returned by download_and_parse_bmrb}

\item{rnaOnly}{logical, write out only RNA chemical shifts?}

\item{listID}{writing chemical shifts corresponding to BMRB list specified by listID}

\item{offset}{add offset residue numbers}

\item{rnaID}{id used in the file name to which chemical shifts are written}

\item{outpath}{path pointing to location to save file}

\item{residues}{if vector not NULL, will only write out chemical shifts specified residues}

\item{nuclei}{if vector not NULL, will only write out chemical shifts specified nuclei}

\item{str}{should chemical shifts be written in STAR format?}

\item{verbose}{logical containing whether info will be printed to screen}
}
\description{
writes out chemcial shift dataframe to file
}
