% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_2D_peaks.R
\name{create_peaks}
\alias{create_peaks}
\title{Peaks Generation Function}
\usage{
create_peaks(cs, type = "hmqc", grouping = c("model", "resid", "resname"),
  protons = c("H1'", "H2'", "H3'", "H4'", "H5'", "H5''", "H2", "H5", "H6",
  "H8"), carbons = c("C1'", "C2'", "C3'", "C4'", "C5'", "C5'", "C2", "C5",
  "C6", "C8"))
}
\arguments{
\item{cs}{input chemical shift dataframe. Should contain field: model, resid, nucleus, weight, and predCS}

\item{type}{type of 2D spectra.}

\item{grouping}{variables used to group data.}

\item{protons}{vector of proton nuclei.}

\item{carbons}{vector of carbon nuclei.}
}
\description{
This function allows you to convert chemical shift list to chemical shift peak table
}
\examples{
create_hmqc_peaks(cs)
}
