% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_library.R, R/utility_functions.R
\name{load_library}
\alias{load_library}
\title{Load Package Function}
\usage{
load_library(package, repo = "http://cran.mtu.edu/")

load_library(package, repo = "http://cran.mtu.edu/")
}
\arguments{
\item{package}{name of package to load.}

\item{repo}{repository (string) to use when installing the package}

\item{package_name}{name (string) of package}
}
\description{
This function loads a package if previous installed or install the package and then loads it.

This function load a package if installed, and install then loads if not
}
\examples{
load_library(package_name='plyr')
}

